pyproject.toml
[project]
name = "pipeline"
version = "0.0.0"
description = "Pre-built template for deep learning pipelines."
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "huggingface-hub>=0.26.1",
    "nanoid>=2.0.0",
    "numpy>=2.1.2",
    "python-dotenv>=1.0.1",
    "rich>=13.9.3",
    "safetensors>=0.4.5",
    "torch>=2.5.0",
    "tqdm>=4.66.5",
    "typer>=0.12.5",
]

[tool.uv]
dev-dependencies = [
    "ipykernel>=6.29.5",
]
